// src/schemas/balancer-api.ts
export interface components {
	schemas: {
// ...
		Route: {
			route: {
				/** @example 1 */
				readonly id?: number;
				/** @example Route */
				name?: string;
				/** @example test.loc.icdc.io */
				hostname?: string;
				/** @example / */
				path?: string;
				/** @example 6587 */
				target_port?: number;
				/** @example None */
				insecure?: string;
				/** @example Edge */
				tls_termination?: string;
				/** @example KV */
				labels?: string;
				/** @example 1 */
				cloud_gateway_id?: number;
				/** @example 1 */
				certificate_id?: number;
				/** @example tcp */
				source_proto?: string;
				/** @example tcp */
				destination_proto?: string;
				/** @example 4 */
				ip_version?: number;
				/** @example user@test.email */
				owner?: string;
				services?: components["schemas"]["Service"][];
				routes_services?: {
					balance_type: string;
					created_at: string;
					id: number;
					name: string;
					route_id: number;
					service_id: number;
					updated_at: string;
					value: number;
				}[];
				/** @example true */
				healthcheck_enabled?: boolean;
				cloud_gateway: {
					account: string;
					cloudgw_id: string;
					cloudgw_instance: string;
					created_at: string;
					id: number;
					name: string;
					updated_at: string;
				};
				healthcheck?: {
					/** @example / */
					path?: string;
					/** @example https */
					scheme?: string;
					/** @example check-cors.loc.icdc.io */
					hostname?: string;
					/** @example 8080 */
					port?: number;
					/** @example 15 */
					interval?: number;
					/** @example 3 */
					timeout?: number;
					/** @example {
					 *       "x-icdc-account": "devel",
					 *       "x-icdc-role": "admin"
					 *     } */
					headers?: Record<string, string>;
					/** @example POST */
					method?: string;
					/** @example true */
					follow_redirects?: boolean;
				};
			};
		};
		Route_POST: {
			route: Omit<components["schemas"]["Route"]["route"], "cloud_gateway">;
		};
	};
// ...
}